GAPUINO8 board:     Clean warnings and some trivia for semihosting support

+++ platform/mcu/gap8/src/semihost.c
@@ -1,4 +1,5 @@
 #include "gap_semihost.h"
+#include "string.h"
 
 // roughly this is the last stage of printf:
 // print a string until '\0'
@@ -10,7 +11,7 @@ void gap8_semihost_write0(const char *print_string)
 int gap8_semihost_open(const char *name, int mode)
 {
     uint32_t len = strlen(name);
-    uint32_t args[3] = {name,mode,len};
+    uint32_t args[3] = {(uint32_t)name,mode,len};
     return __internal_semihost(SEMIHOSTING_SYS_OPEN, (long) args);
 }
